+++ src/infra/core/src/pull_service_impl.rs
@@ -269,7 +269,7 @@ impl PullServiceImpl {
         use tokio_stream::StreamExt;
         let mut datasets = self
             .dataset_repo
-            .get_account_datasets(self.current_account_subject.account_name.clone());
+            .get_datasets_by_owner(self.current_account_subject.account_name.clone());
         while let Some(dataset_handle) = datasets.next().await {
             let dataset_handle = dataset_handle?;