+++ src/infra/core/tests/tests/engine/test_engine_io.rs
@@ -245,6 +245,7 @@ async fn test_engine_io_s3_to_local_file_mount_proxy() {
     let dataset_repo = Arc::new(DatasetRepositoryS3::new(
         s3_context,
         Arc::new(CurrentAccountSubject::new_test()),
+        false,
     ));
 
     test_engine_io_common(