+++ src/infra/core/tests/tests/repos/test_dataset_repository_s3.rs
@@ -9,18 +9,23 @@
 
 use std::sync::Arc;
 
+use kamu::domain::{CurrentAccountSubject, TEST_ACCOUNT_NAME};
 use kamu::testing::LocalS3Server;
 use kamu::utils::s3_context::S3Context;
 use kamu::DatasetRepositoryS3;
-use kamu_core::CurrentAccountSubject;
+use opendatafabric::AccountName;
 
 use super::test_dataset_repository_shared;
 
 /////////////////////////////////////////////////////////////////////////////////////////
 
-async fn s3_repo(s3: &LocalS3Server) -> DatasetRepositoryS3 {
+async fn s3_repo(s3: &LocalS3Server, multi_tenant: bool) -> DatasetRepositoryS3 {
     let s3_context = S3Context::from_url(&s3.url).await;
-    DatasetRepositoryS3::new(s3_context, Arc::new(CurrentAccountSubject::new_test()))
+    DatasetRepositoryS3::new(
+        s3_context,
+        Arc::new(CurrentAccountSubject::new_test()),
+        multi_tenant,
+    )
 }
 
 /////////////////////////////////////////////////////////////////////////////////////////
@@ -29,7 +34,7 @@ async fn s3_repo(s3: &LocalS3Server) -> DatasetRepositoryS3 {
 #[tokio::test]
 async fn test_create_dataset() {
     let s3 = LocalS3Server::new().await;
-    let repo = s3_repo(&s3).await;
+    let repo = s3_repo(&s3, false).await;
 
     test_dataset_repository_shared::test_create_dataset(&repo, None).await;
 }
@@ -38,9 +43,36 @@ async fn test_create_dataset() {
 
 #[test_group::group(containerized)]
 #[tokio::test]
+async fn test_create_dataset_multi_tenant() {
+    let s3 = LocalS3Server::new().await;
+    let repo = s3_repo(&s3, true).await;
+
+    test_dataset_repository_shared::test_create_dataset(
+        &repo,
+        Some(AccountName::new_unchecked(TEST_ACCOUNT_NAME)),
+    )
+    .await;
+}
+
+/////////////////////////////////////////////////////////////////////////////////////////
+
+#[test_group::group(containerized)]
+#[tokio::test]
+async fn test_create_dataset_same_name_multiple_tenants() {
+    let s3 = LocalS3Server::new().await;
+    let repo = s3_repo(&s3, true).await;
+
+    test_dataset_repository_shared::test_create_dataset_same_name_multiple_tenants(&repo).await;
+}
+
+/////////////////////////////////////////////////////////////////////////////////////////
+
+#[test_group::group(containerized)]
+#[test_group::group(containerized)]
+#[tokio::test]
 async fn test_create_dataset_from_snapshot() {
     let s3 = LocalS3Server::new().await;
-    let repo = s3_repo(&s3).await;
+    let repo = s3_repo(&s3, false).await;
 
     test_dataset_repository_shared::test_create_dataset_from_snapshot(&repo, None).await;
 }
@@ -49,9 +81,24 @@ async fn test_create_dataset_from_snapshot() {
 
 #[test_group::group(containerized)]
 #[tokio::test]
+async fn test_create_dataset_from_snapshot_multi_tenant() {
+    let s3 = LocalS3Server::new().await;
+    let repo = s3_repo(&s3, true).await;
+
+    test_dataset_repository_shared::test_create_dataset_from_snapshot(
+        &repo,
+        Some(AccountName::new_unchecked(TEST_ACCOUNT_NAME)),
+    )
+    .await;
+}
+
+/////////////////////////////////////////////////////////////////////////////////////////
+
+#[test_group::group(containerized)]
+#[tokio::test]
 async fn test_rename_dataset() {
     let s3 = LocalS3Server::new().await;
-    let repo = s3_repo(&s3).await;
+    let repo = s3_repo(&s3, false).await;
 
     test_dataset_repository_shared::test_rename_dataset(&repo, None).await;
 }
@@ -60,9 +107,35 @@ async fn test_rename_dataset() {
 
 #[test_group::group(containerized)]
 #[tokio::test]
+async fn test_rename_dataset_multi_tenant() {
+    let s3 = LocalS3Server::new().await;
+    let repo = s3_repo(&s3, true).await;
+
+    test_dataset_repository_shared::test_rename_dataset(
+        &repo,
+        Some(AccountName::new_unchecked(TEST_ACCOUNT_NAME)),
+    )
+    .await;
+}
+
+/////////////////////////////////////////////////////////////////////////////////////////
+
+#[test_group::group(containerized)]
+#[tokio::test]
+async fn test_rename_dataset_same_name_multiple_tenants() {
+    let s3 = LocalS3Server::new().await;
+    let repo = s3_repo(&s3, true).await;
+
+    test_dataset_repository_shared::test_rename_dataset_same_name_multiple_tenants(&repo).await;
+}
+
+/////////////////////////////////////////////////////////////////////////////////////////
+
+#[test_group::group(containerized)]
+#[tokio::test]
 async fn test_delete_dataset() {
     let s3 = LocalS3Server::new().await;
-    let repo = s3_repo(&s3).await;
+    let repo = s3_repo(&s3, false).await;
 
     test_dataset_repository_shared::test_delete_dataset(&repo, None).await;
 }
@@ -71,11 +144,37 @@ async fn test_delete_dataset() {
 
 #[test_group::group(containerized)]
 #[tokio::test]
+async fn test_delete_dataset_multi_tenant() {
+    let s3 = LocalS3Server::new().await;
+    let repo = s3_repo(&s3, true).await;
+
+    test_dataset_repository_shared::test_delete_dataset(
+        &repo,
+        Some(AccountName::new_unchecked(TEST_ACCOUNT_NAME)),
+    )
+    .await;
+}
+
+/////////////////////////////////////////////////////////////////////////////////////////
+
+#[test_group::group(containerized)]
+#[tokio::test]
 async fn test_iterate_datasets() {
     let s3 = LocalS3Server::new().await;
-    let repo = s3_repo(&s3).await;
+    let repo = s3_repo(&s3, false).await;
 
     test_dataset_repository_shared::test_iterate_datasets(&repo).await;
+}
+
+/////////////////////////////////////////////////////////////////////////////////////////
+
+#[test_group::group(containerized)]
+#[tokio::test]
+async fn test_iterate_datasets_multi_tenant() {
+    let s3 = LocalS3Server::new().await;
+    let repo = s3_repo(&s3, true).await;
+
+    test_dataset_repository_shared::test_iterate_datasets_multi_tenant(&repo).await;
 }
 
 /////////////////////////////////////////////////////////////////////////////////////////