+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2023-08-03
+### Added
+- Support multi-tenancy within S3-based dataset repository
+
 ## [0.135.0] - 2023-08-01
 ### Added
 - New `kamu version` command that outputs detailed build information in JSON or YAML